backend/requirements.txt
fastapi==0.117.1
uvicorn==0.22.0
numpy==1.25.2
opencv-python==4.8.1.78
scikit-learn==1.3.2
tensorflow==2.14.0
mediapipe==0.10.13
pydantic==2.6.1
google-generativeai==0.0.2
